+++ src/app/group/[groupId]/page.tsx
@@ -1,11 +1,13 @@
 import { api } from '@/trpc/server'
-// import { generateTableData } from '@/utils/generateTableData'
+import { generateTableData } from '@/utils/generateTableData'
 import { redis } from '@/config/redis'
 import type { GroupData } from '@/types/tableRow.type'
 import type { roadmap } from '@prisma/client'
+import { NOT_STARTED_GROUPS } from '@/data/constants'
 
-// import { TrackTable } from '@/app/group/_components/track-table'
-// import { ConfettiFireworks } from '@/app/group/_components/confetti-fireworks'
+import { TrackTable } from '@/app/group/_components/track-table'
+import { ConfettiFireworks } from '@/app/group/_components/confetti-fireworks'
+import { NotStarted } from '@/app/group/_components/not-started'
 
 type CachedData = {
   groupData: GroupData
@@ -14,6 +16,9 @@ type CachedData = {
 
 const Page = async ({ params }: { params: Promise<{ groupId: string }> }) => {
   const { groupId } = await params
+
+  // for not started groups
+  if (NOT_STARTED_GROUPS.includes(+groupId)) return <NotStarted />
 
   const user = await api.auth.getUser()
   console.log('user?.user_metadata', user?.user_metadata)
@@ -34,39 +39,30 @@ const Page = async ({ params }: { params: Promise<{ groupId: string }> }) => {
     roadmap = await api.roadmap.getGroupProblems(groupData?.group_progress || [])
 
     if (groupData && roadmap) {
-      await redis.set(cacheKey, { groupData, roadmap }, { ex: 300 }) // cache for
+      await redis.set(cacheKey, { groupData, roadmap }, { ex: 180 }) // cache for 3 minutes
     }
   }
 
-  // if (!groupData || !roadmap || roadmap.length === 0) {
+  if (!groupData || !roadmap || roadmap.length === 0) return <NotStarted />
+
+  const tableData = generateTableData({
+    group_no: groupData.group_no,
+    submission: groupData.submissions,
+    roadmap,
+    group_progress: groupData.group_progress,
+  })
+
   return (
-    <div className="flex flex-1 flex-col items-center justify-center gap-3 px-4">
-      <h1 className="max-w-2xl text-center text-lg font-semibold md:text-2xl">
-        Get ready! Your group&apos;s journey is being prepared. We will notify you when it&apos;s
-        ready. 🔥
-      </h1>
-    </div>
+    <>
+      <TrackTable
+        leetcoders={groupData.leetcoders}
+        tableData={tableData}
+        groupId={groupId}
+      />
+
+      <ConfettiFireworks />
+    </>
   )
-  // }
-
-  // const tableData = generateTableData({
-  //   group_no: groupData.group_no,
-  //   submission: groupData.submissions,
-  //   roadmap,
-  //   group_progress: groupData.group_progress,
-  // })
-
-  // return (
-  //   <>
-  //     <TrackTable
-  //       leetcoders={groupData.leetcoders}
-  //       tableData={tableData}
-  //       groupId={groupId}
-  //     />
-
-  //     <ConfettiFireworks />
-  //   </>
-  // )
 }
 
 export default Page