+++ CHANGELOG.md
@@ -1,28 +1,106 @@
-### Changelog
+# Changelog
 
-- 2.0.0: FEATURE: Rebuild the application with modern technologies, features and more & organized & reliable structure.
-- 1.2.22: FEAT: add logout icon
-- 1.2.21: FEAT: add leetcoder card
-- 1.1.21: FIX: authorization issue
-- 1.1.20: ENHANCEMENT: make logo as Link instead of image
-- 1.1.19: FIX: update fetching groupProgress logic
-- 1.1.18: FEAT: add update profile page
-- 1.0.18: FEAT: use [neetcode](https://neetcode.io/practice) roadmap
-- 1.0.17: FIX: use gmail instead of titan
-- 1.0.16: STYLE: update roadmap accordion style
-- 1.0.15: FEAT: add `/roadmap` page
-- 1.0.14: FEAT: configure Redis
-- 1.0.13: ENHANCEMENT: use nodemailer instead of resend
-- 1.0.12: STYLE: update count chart
-- 1.0.11: STYLE: update scrollbar style
-- 1.0.10: FEAT: add navigation menu
-- 1.0.9: STYLE: update scrollbar style
-- 1.0.8: FEAT: add chart for count
-- 1.0.7: FEAT: add group 2
-- 1.0.6: FEAT: kick off neglected leetcoders after 5 non-solved problems
-- 1.0.5: FEAT: secure cron job
-- 1.0.4: ENHANCEMENT: use [cron-jon](https://console.cron-job.org/jobs) instead of vercel cron
-- 1.0.3: FEAT: automate "send daily problem email" cron job
-- 1.0.2: FEAT: automate "add group progress" cron job
-- 1.0.1: FIX: show validation error message when the form is submitted
-- 1.0.0: Initial release
+## [2.0.9]
+fix adding group progress for not started groups
+
+## [2.0.8]
+update toast styles
+
+## [2.0.7]
+backup each 1 hour
+
+## [2.0.6]
+update group cards style
+
+## [2.0.5]
+fix group progress query bug
+
+## [2.0.4]
+update the style of checkbox
+
+## [2.0.3]
+fix email daily problem email button
+
+## [2.0.2]
+send admin notifications from the server for better performance
+
+## [2.0.1]
+if the user is not logged in, redirect to the login page
+
+## [2.0.0]
+Rebuild the application with modern technologies, features and more & organized & reliable structure.
+
+## [1.2.22]
+add logout icon
+
+## [1.2.21]
+add leetcoder card
+
+## [1.1.21]
+authorization issue
+
+## [1.1.20]
+ENHANCEMENT: make logo as Link instead of image
+
+## [1.1.19]
+update fetching groupProgress logic
+
+## [1.1.18]
+add update profile page
+
+## [1.0.18]
+use [neetcode](https://neetcode.io/practice) roadmap
+
+## [1.0.17]
+use gmail instead of titan
+
+## [1.0.16]
+STYLE: update roadmap accordion style
+
+## [1.0.15]
+add `/roadmap` page
+
+## [1.0.14]
+configure Redis
+
+## [1.0.13]
+ENHANCEMENT: use nodemailer instead of resend
+
+## [1.0.12]
+STYLE: update count chart
+
+## [1.0.11]
+STYLE: update scrollbar style
+
+## [1.0.10]
+add navigation menu
+
+## [1.0.9]
+STYLE: update scrollbar style
+
+## [1.0.8]
+add chart for count
+
+## [1.0.7]
+add group 2
+
+## [1.0.6]
+kick off neglected leetcoders after 5 non-solved problems
+
+## [1.0.5]
+secure cron job
+
+## [1.0.4]
+ENHANCEMENT: use [cron-jon](https://console.cron-job.org/jobs) instead of vercel cron
+
+## [1.0.3]
+automate "send daily problem email" cron job
+
+## [1.0.2]
+automate "add group progress" cron job
+
+## [1.0.1]
+show validation error message when the form is submitted
+
+## [1.0.0]
+Initial release